requirements.txt
langchain
langchain-openai
langchain-anthropic
python-dotenv
datasets
tqdm
rouge
matplotlib